requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.4
tzdata==2025.2
python-dotenv==1.0.1
gunicorn==23.0.0
aiohttp==3.13.2
asyncio==4.0.0